predict with conjugate Normal-Inverse-Wishart bayesian VAR model

predict with conjugate Normal-Inverse-Wishart bayesian VAR model

bvar_conj_forecast(model, Y_in = NULL, Z_f = NULL, output = c("long",
  "wide"), h = 1, out_of_sample = TRUE, type = c("prediction",
  "credible"), level = c(80, 95), include = c("mean", "median", "sd",
  "interval", "raw"), fast_forecast = FALSE, verbose = FALSE)

Arguments

model
estimated conjugate N-IW model

Y_in
(NULL by default) past values of endogeneous variables (shold have at least p observations). If NULL, then Y_in supplied for estimation will be used. For out-of-sample forecast only last p values of Y_in are used

Z_f
future values of exogeneous variables

output
(default 'long') --- long or wide table for mean/quantiles of forecasts

h
number of periods for forecasting

out_of_sample
logical, default is TRUE, whether forecasts are out-of-sample or in-sample. If forecasts are not out-of-sample, then parameter h is ignored

type
('prediction' by default) type of interval: 'prediction' incorporates uncertainty about future shocks; 'credible' deals only with parameter uncertainty.

level
confidence levels for prediction intervals

include
(default is c('mean', 'median', 'sd', 'raw')) what type of summary to provide If 'raw' is present raw forecasts will be reported. If only 'raw' is present then function will return coda mcmc object with raw forecasts. Otherwise raw forecasts will be attached as attribute.

fast_forecast
logical, FALSE by default. If TRUE then only mean forecast is calculated, posterior expected values of hyperparameters are used. No confidence intervals, no sd, no median. This mode is activated by default if there are no simulations in supplied model.

verbose
(default FALSE) if true some messages will be printed

Value

forecast results

Details

predict with conjugate Normal-Inverse-Wishart bayesian VAR model

Examples

data(Yraw)
setup <- bvar_conj_setup(Yraw, p = 4)

#> Calculation of hyperparameters from lambdas is not implemented yet :(

#> You may safely ignore the message if you supply lambdas :)

model <- bvar_conj_estimate(setup, keep = 100)
bvar_conj_forecast(model, h=2, output = 'wide')

#>        variable h     mean   median        sd lower_80 upper_80 lower_95
#> 1     inflation 1 2.832796 2.814138 0.3403388 2.386363 3.310531 2.192906
#> 2     inflation 2 2.850935 2.834378 0.5268631 2.162197 3.592316 1.959676
#> 3 interest_rate 1 4.725440 4.769359 0.7426514 3.825763 5.632314 3.382020
#> 4 interest_rate 2 4.633839 4.671352 0.9760099 3.331887 5.837528 2.821528
#> 5  unemployment 1 4.875626 4.872675 0.3145944 4.494299 5.285674 4.302123
#> 6  unemployment 2 5.006729 4.997023 0.4861070 4.364303 5.581539 4.236558
#>   upper_95
#> 1 3.531297
#> 2 3.890123
#> 3 6.231145
#> 4 6.198599
#> 5 5.476999
#> 6 6.019809

bvar_conj_forecast(model, out_of_sample = FALSE, include = 'mean')

#>          variable   h what      value
#> 1       inflation   1 mean  1.0761624
#> 2    unemployment   1 mean  3.3854080
#> 3   interest_rate   1 mean  1.2185083
#> 4       inflation   2 mean  1.1504176
#> 5    unemployment   2 mean  5.4294020
#> 6   interest_rate   2 mean  0.4049258
#> 7       inflation   3 mean  0.9984510
#> 8    unemployment   3 mean  6.1410399
#> 9   interest_rate   3 mean  0.4521094
#> 10      inflation   4 mean  0.7504793
#> 11   unemployment   4 mean  5.6641571
#> 12  interest_rate   4 mean  0.6751586
#> 13      inflation   5 mean  0.6431546
#> 14   unemployment   5 mean  5.4078612
#> 15  interest_rate   5 mean  1.1380173
#> 16      inflation   6 mean  0.8659616
#> 17   unemployment   6 mean  4.5207253
#> 18  interest_rate   6 mean  1.7216726
#> 19      inflation   7 mean  1.3709615
#> 20   unemployment   7 mean  4.4185772
#> 21  interest_rate   7 mean  1.9408150
#> 22      inflation   8 mean  2.4289895
#> 23   unemployment   8 mean  3.7671262
#> 24  interest_rate   8 mean  2.4206587
#> 25      inflation   9 mean  2.7502064
#> 26   unemployment   9 mean  4.3083798
#> 27  interest_rate   9 mean  2.4640353
#> 28      inflation  10 mean  3.3965147
#> 29   unemployment  10 mean  4.0105496
#> 30  interest_rate  10 mean  2.9294784
#> 31      inflation  11 mean  3.7051646
#> 32   unemployment  11 mean  4.0421181
#> 33  interest_rate  11 mean  3.1044357
#> 34      inflation  12 mean  4.0024652
#> 35   unemployment  12 mean  4.6362592
#> 36  interest_rate  12 mean  2.4878805
#> 37      inflation  13 mean  3.8773179
#> 38   unemployment  13 mean  3.8806465
#> 39  interest_rate  13 mean  3.4216847
#> 40      inflation  14 mean  4.0672832
#> 41   unemployment  14 mean  4.3371133
#> 42  interest_rate  14 mean  3.3495523
#> 43      inflation  15 mean  3.7576980
#> 44   unemployment  15 mean  3.9535982
#> 45  interest_rate  15 mean  3.3265851
#> 46      inflation  16 mean  3.1544958
#> 47   unemployment  16 mean  4.4194207
#> 48  interest_rate  16 mean  2.9620054
#> 49      inflation  17 mean  3.1301655
#> 50   unemployment  17 mean  4.7722540
#> 51  interest_rate  17 mean  3.2999707
#> 52      inflation  18 mean  2.6644854
#> 53   unemployment  18 mean  6.1533070
#> 54  interest_rate  18 mean  1.9170518
#> 55      inflation  19 mean  2.1338676
#> 56   unemployment  19 mean  7.7854385
#> 57  interest_rate  19 mean  0.6068460
#> 58      inflation  20 mean  1.6506579
#> 59   unemployment  20 mean  7.3799461
#> 60  interest_rate  20 mean  0.8367000
#> 61      inflation  21 mean  1.5436005
#> 62   unemployment  21 mean  6.1982303
#> 63  interest_rate  21 mean  2.6812697
#> 64      inflation  22 mean  1.3272332
#> 65   unemployment  22 mean  5.4873491
#> 66  interest_rate  22 mean  2.9436177
#> 67      inflation  23 mean  1.1396353
#> 68   unemployment  23 mean  4.8325633
#> 69  interest_rate  23 mean  3.2846762
#> 70      inflation  24 mean  1.1513869
#> 71   unemployment  24 mean  4.9089487
#> 72  interest_rate  24 mean  3.2726274
#> 73      inflation  25 mean  1.2955917
#> 74   unemployment  25 mean  5.7056003
#> 75  interest_rate  25 mean  3.7450702
#> 76      inflation  26 mean  1.0587608
#> 77   unemployment  26 mean  5.0040394
#> 78  interest_rate  26 mean  4.1651518
#> 79      inflation  27 mean  1.5330326
#> 80   unemployment  27 mean  5.1747816
#> 81  interest_rate  27 mean  3.3908141
#> 82      inflation  28 mean  1.6757754
#> 83   unemployment  28 mean  5.5661108
#> 84  interest_rate  28 mean  2.6402677
#> 85      inflation  29 mean  1.4554466
#> 86   unemployment  29 mean  6.1404854
#> 87  interest_rate  29 mean  2.2700141
#> 88      inflation  30 mean  1.3160169
#> 89   unemployment  30 mean  6.5637418
#> 90  interest_rate  30 mean  1.9232645
#> 91      inflation  31 mean  0.9758710
#> 92   unemployment  31 mean  6.8672722
#> 93  interest_rate  31 mean  2.1221360
#> 94      inflation  32 mean  0.8082238
#> 95   unemployment  32 mean  6.7770021
#> 96  interest_rate  32 mean  2.1538883
#> 97      inflation  33 mean  0.7839792
#> 98   unemployment  33 mean  6.2027170
#> 99  interest_rate  33 mean  2.4160207
#> 100     inflation  34 mean  1.5511278
#> 101  unemployment  34 mean  5.3823974
#> 102 interest_rate  34 mean  3.1052012
#> 103     inflation  35 mean  1.5785798
#> 104  unemployment  35 mean  5.3258156
#> 105 interest_rate  35 mean  2.9642810
#> 106     inflation  36 mean  1.4974885
#> 107  unemployment  36 mean  5.1834387
#> 108 interest_rate  36 mean  3.2524976
#> 109     inflation  37 mean  1.5444969
#> 110  unemployment  37 mean  5.3109683
#> 111 interest_rate  37 mean  2.9471932
#> 112     inflation  38 mean  0.9475982
#> 113  unemployment  38 mean  5.6765520
#> 114 interest_rate  38 mean  2.7383104
#> 115     inflation  39 mean  0.9393843
#> 116  unemployment  39 mean  5.6308592
#> 117 interest_rate  39 mean  2.6638137
#> 118     inflation  40 mean  0.9248611
#> 119  unemployment  40 mean  5.5002926
#> 120 interest_rate  40 mean  2.9491123
#> 121     inflation  41 mean  1.3283138
#> 122  unemployment  41 mean  5.3428867
#> 123 interest_rate  41 mean  3.3310130
#> 124     inflation  42 mean  1.3615126
#> 125  unemployment  42 mean  5.4904848
#> 126 interest_rate  42 mean  3.5686751
#> 127     inflation  43 mean  1.6040403
#> 128  unemployment  43 mean  5.1784027
#> 129 interest_rate  43 mean  3.5611347
#> 130     inflation  44 mean  2.0470118
#> 131  unemployment  44 mean  4.6978189
#> 132 interest_rate  44 mean  3.7890262
#> 133     inflation  45 mean  1.6610518
#> 134  unemployment  45 mean  5.1537229
#> 135 interest_rate  45 mean  3.6746383
#> 136     inflation  46 mean  1.7572455
#> 137  unemployment  46 mean  4.8429004
#> 138 interest_rate  46 mean  3.7119253
#> 139     inflation  47 mean  1.9945369
#> 140  unemployment  47 mean  4.8084761
#> 141 interest_rate  47 mean  3.7887140
#> 142     inflation  48 mean  1.9754493
#> 143  unemployment  48 mean  4.3045246
#> 144 interest_rate  48 mean  4.1000724
#> 145     inflation  49 mean  2.1532411
#> 146  unemployment  49 mean  4.1864797
#> 147 interest_rate  49 mean  4.0893966
#> 148     inflation  50 mean  2.3483502
#> 149  unemployment  50 mean  4.0820615
#> 150 interest_rate  50 mean  4.5953809
#> 151     inflation  51 mean  3.0189105
#> 152  unemployment  51 mean  3.8338778
#> 153 interest_rate  51 mean  4.8419464
#> 154     inflation  52 mean  3.5521912
#> 155  unemployment  52 mean  3.9395878
#> 156 interest_rate  52 mean  5.0424432
#> 157     inflation  53 mean  3.7163715
#> 158  unemployment  53 mean  3.8246161
#> 159 interest_rate  53 mean  5.4144232
#> 160     inflation  54 mean  3.4302690
#> 161  unemployment  54 mean  4.1435444
#> 162 interest_rate  54 mean  4.7398599
#> 163     inflation  55 mean  2.9831743
#> 164  unemployment  55 mean  4.0041809
#> 165 interest_rate  55 mean  3.9971311
#> 166     inflation  56 mean  3.0590378
#> 167  unemployment  56 mean  3.9803350
#> 168 interest_rate  56 mean  4.2011675
#> 169     inflation  57 mean  3.3244387
#> 170  unemployment  57 mean  4.2680340
#> 171 interest_rate  57 mean  4.4567635
#> 172     inflation  58 mean  4.1421717
#> 173  unemployment  58 mean  3.8610350
#> 174 interest_rate  58 mean  5.2345976
#> 175     inflation  59 mean  4.6787915
#> 176  unemployment  59 mean  3.6896333
#> 177 interest_rate  59 mean  5.7024184
#> 178     inflation  60 mean  4.5291577
#> 179  unemployment  60 mean  3.9472456
#> 180 interest_rate  60 mean  5.3405876
#> 181     inflation  61 mean  4.8652082
#> 182  unemployment  61 mean  3.6607629
#> 183 interest_rate  61 mean  5.6141995
#> 184     inflation  62 mean  4.7198132
#> 185  unemployment  62 mean  3.6119236
#> 186 interest_rate  62 mean  6.2119051
#> 187     inflation  63 mean  5.0108320
#> 188  unemployment  63 mean  3.7116266
#> 189 interest_rate  63 mean  6.1843917
#> 190     inflation  64 mean  5.6396585
#> 191  unemployment  64 mean  3.8387660
#> 192 interest_rate  64 mean  7.0575508
#> 193     inflation  65 mean  5.1919325
#> 194  unemployment  65 mean  4.0537643
#> 195 interest_rate  65 mean  6.7512086
#> 196     inflation  66 mean  5.7693489
#> 197  unemployment  66 mean  4.1910940
#> 198 interest_rate  66 mean  7.8614331
#> 199     inflation  67 mean  5.6918215
#> 200  unemployment  67 mean  5.1015513
#> 201 interest_rate  67 mean  6.3371810
#> 202     inflation  68 mean  4.7017117
#> 203  unemployment  68 mean  5.3965389
#> 204 interest_rate  68 mean  6.1096325
#> 205     inflation  69 mean  4.8255561
#> 206  unemployment  69 mean  5.7830113
#> 207 interest_rate  69 mean  5.6403799
#> 208     inflation  70 mean  5.0614981
#> 209  unemployment  70 mean  6.2336944
#> 210 interest_rate  70 mean  4.4294614
#> 211     inflation  71 mean  4.9570246
#> 212  unemployment  71 mean  6.0994272
#> 213 interest_rate  71 mean  4.1602183
#> 214     inflation  72 mean  5.2492899
#> 215  unemployment  72 mean  5.9568811
#> 216 interest_rate  72 mean  5.4874760
#> 217     inflation  73 mean  4.5891065
#> 218  unemployment  73 mean  5.7954945
#> 219 interest_rate  73 mean  4.4674498
#> 220     inflation  74 mean  4.9339999
#> 221  unemployment  74 mean  5.8714722
#> 222 interest_rate  74 mean  3.8790080
#> 223     inflation  75 mean  3.8615889
#> 224  unemployment  75 mean  5.6821733
#> 225 interest_rate  75 mean  3.7414245
#> 226     inflation  76 mean  3.9482525
#> 227  unemployment  76 mean  5.5885391
#> 228 interest_rate  76 mean  4.1338061
#> 229     inflation  77 mean  4.5508455
#> 230  unemployment  77 mean  5.6506717
#> 231 interest_rate  77 mean  4.8653303
#> 232     inflation  78 mean  4.1901267
#> 233  unemployment  78 mean  4.7660417
#> 234 interest_rate  78 mean  5.6216200
#> 235     inflation  79 mean  5.5128305
#> 236  unemployment  79 mean  4.9846673
#> 237 interest_rate  79 mean  6.6688686
#> 238     inflation  80 mean  6.7189919
#> 239  unemployment  80 mean  4.8609325
#> 240 interest_rate  80 mean  8.2150574
#> 241     inflation  81 mean  7.2373019
#> 242  unemployment  81 mean  4.7130031
#> 243 interest_rate  81 mean  7.7001168
#> 244     inflation  82 mean  7.8837137
#> 245  unemployment  82 mean  5.4039276
#> 246 interest_rate  82 mean  8.1527062
#> 247     inflation  83 mean  8.3836916
#> 248  unemployment  83 mean  5.3854527
#> 249 interest_rate  83 mean  8.6083828
#> 250     inflation  84 mean  9.8655787
#>  [ reached getOption("max.print") -- omitted 383 rows ]