summary of a conjugate Normal-Inverse-Wishart bayesian VAR model

summary of a conjugate Normal-Inverse-Wishart bayesian VAR model

bvar_conj_summary(model)

Arguments

model
estimated conjugate N-IW model

Value

nothing

Details

summary of a conjugate Normal-Inverse-Wishart bayesian VAR model

Examples

data(Yraw)
setup <- bvar_conj_setup(Yraw, p = 4)

#> Calculation of hyperparameters from lambdas is not implemented yet :(

#> You may safely ignore the message if you supply lambdas :)

model <- bvar_conj_estimate(setup = setup, keep = 100)
bvar_conj_summary(model)

#> Number of lags, p = 4

#> Number of endogeneos variables, m = 3

#> Number of exogeneos variables (including constant), d = 1

#> Number of parameters, k = mp + d = 13

#> Initial number of observations, T_in = 215

#> Number of dummy observations, T_dummy = 20

#> Number of observations available for classic VAR, T = T_in - p = 211

#> Posterior mean of Phi (VAR coefficients) [k = 13 x m = 3]:

#>                      eq_inflation eq_unemployment eq_interest_rate
#> inflation, l = 1      1.324303477     0.026892371      0.326364798
#> unemployment, l = 1  -0.192550805     1.171699260     -0.429377021
#> interest_rate, l = 1 -0.026508511    -0.029864388      0.832926444
#> inflation, l = 2     -0.221426553    -0.011118977     -0.178665196
#> unemployment, l = 2   0.065438617    -0.199306705      0.303542308
#> interest_rate, l = 2  0.019102471     0.016428946     -0.003814684
#> inflation, l = 3     -0.085212104    -0.015393879      0.051025071
#> unemployment, l = 3   0.041503852    -0.087072758      0.053237955
#> interest_rate, l = 3  0.004068285     0.036506618      0.060082575
#> inflation, l = 4     -0.020719579     0.030347152     -0.098299318
#> unemployment, l = 4   0.038393564     0.003018880      0.092319262
#> interest_rate, l = 4 -0.001007131     0.005102153      0.012306069
#> const                 0.313102595     0.396227579      0.055204953

#> Posterior nu = 216

#> Number of mcmc simulations, keep = 100

#> Posterior sample mean of Phi (VAR coefficients) [k = 13 x m = 3]:

#>                      eq_inflation eq_unemployment eq_interest_rate
#> inflation, l = 1      1.325249881     0.035847898      0.315937276
#> unemployment, l = 1  -0.186128512     1.185347745     -0.451734641
#> interest_rate, l = 1 -0.026340672    -0.030025586      0.827765355
#> inflation, l = 2     -0.224780213    -0.026596304     -0.157063837
#> unemployment, l = 2   0.059430562    -0.218134370      0.319779275
#> interest_rate, l = 2  0.020767942     0.016620542     -0.008443557
#> inflation, l = 3     -0.079560356    -0.013148946      0.051628900
#> unemployment, l = 3   0.041095031    -0.081167773      0.053341139
#> interest_rate, l = 3  0.003558498     0.035013170      0.075663368
#> inflation, l = 4     -0.025323754     0.035340311     -0.107709407
#> unemployment, l = 4   0.039481458     0.002805226      0.088423957
#> interest_rate, l = 4 -0.001577590     0.005577362      0.007477583
#> const                 0.308482905     0.396161448      0.099250522

#> Posterior sample mean of Sigma (noise covariance matrix) [m = 3 x m = 3]:

#>                  inflation unemployment interest_rate
#> inflation      0.101074530 -0.005353771    0.03636750
#> unemployment  -0.005353771  0.108798865   -0.09683692
#> interest_rate  0.036367498 -0.096836920    0.59015487

#> Posterior sample sd of Phi (VAR coefficients) [k = 13 x m = 3]:

#>                      eq_inflation eq_unemployment eq_interest_rate
#> inflation, l = 1       0.05535204      0.05692719       0.12629234
#> unemployment, l = 1    0.05408410      0.05519552       0.13007327
#> interest_rate, l = 1   0.02261849      0.02081705       0.06406205
#> inflation, l = 2       0.06703651      0.07595641       0.14793185
#> unemployment, l = 2    0.05643154      0.06447660       0.15547279
#> interest_rate, l = 2   0.02251012      0.02256099       0.06182725
#> inflation, l = 3       0.04807926      0.04818159       0.12096756
#> unemployment, l = 3    0.04526679      0.04192627       0.09912910
#> interest_rate, l = 3   0.01879111      0.02135144       0.04626154
#> inflation, l = 4       0.03317886      0.03577585       0.07393354
#> unemployment, l = 4    0.03153718      0.03473662       0.07507743
#> interest_rate, l = 4   0.01423018      0.01615086       0.03912260
#> const                  0.09352182      0.10263803       0.26048028

#> Posterior sample sd of Sigma (noise covariance matrix) [m = 3 x m = 3]:

#>                 inflation unemployment interest_rate
#> inflation     0.008734558  0.006507812    0.01855960
#> unemployment  0.006507812  0.011456540    0.01867098
#> interest_rate 0.018559603  0.018670975    0.06372792