Estimate conjugate Normal-Inverse-Wishart bayesian VAR model

Estimate conjugate Normal-Inverse-Wishart bayesian VAR model

bvar_conjugate0(Y_in = NULL, Z_in = NULL, constant = TRUE, p = NULL,
  keep = 10000, verbose = FALSE, priors = list(), fast_forecast = FALSE,
  way_omega_post_root = c("cholesky", "svd"))

Arguments

Y_in
the matrix or data.frame with endogeneous VAR variables

Z_in
(NULL by default) the matrix or data.frame with exogeneous VAR variables

constant
(TRUE by default) whether we should include constant

p
(2 by default) the number of lags

keep
(10000 by default) the number of Gibbs sampling replications to keep Is ignored when the fast_forecast is TRUE.

verbose
(FALSE by default)

priors
the list containing at least Phi_prior [k x m], Omega_prior [k x k], S_prior [m x m], v_prior [1x1], it may also contain Y_dummy [T_dummy x m], X_dummy [T_dummy x k] where k = mp+d

fast_forecast
logical, FALSE by default. If TRUE then no simulations are done, only posterior hyperparameters are calculated.

way_omega_post_root
the way for (Omega_post)^1/2 calculation: 'svd' or 'cholesky'

Value

the list containing all results of bayesian VAR estimation

Details

Estimate conjugate Normal-Inverse-Wishart bayesian VAR model

Examples

data(Yraw)
priors <- Carriero_priors(Yraw, p = 4)
model <- bvar_conjugate0(priors = priors, keep = 100)